# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory": {
      "get": {
        "operationId": "InventoryController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List inventory with product details",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/{id}/reservations": {
      "post": {
        "operationId": "InventoryController_reserve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reserve stock atomically",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/{id}/releases": {
      "post": {
        "operationId": "InventoryController_release",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Release reserved stock atomically",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/{id}/adjustments": {
      "post": {
        "operationId": "InventoryController_adjust",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adjust physical stock with a reason ledger",
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/promotions": {
      "get": {
        "operationId": "PromotionsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List promotions",
        "tags": [
          "Promotions"
        ]
      },
      "post": {
        "operationId": "PromotionsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an HFSS-aware promotion",
        "tags": [
          "Promotions"
        ]
      }
    },
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "CustomerController_reviews",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List approved product reviews",
        "tags": [
          "Customer account"
        ]
      },
      "post": {
        "operationId": "CustomerController_createReview",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer account"
        ]
      }
    },
...
    "/api/v1/admin/dashboard": {
      "get": {
        "operationId": "AdminController_dashboard",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/reports/sales": {
      "get": {
        "operationId": "AdminController_sales",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/reports/customers": {
      "get": {
        "operationId": "AdminController_customerReport",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/reports/products": {
      "get": {
        "operationId": "AdminController_productReport",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/reports/coupons": {
      "get": {
        "operationId": "AdminController_couponReport",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/customers": {
      "get": {
        "operationId": "AdminController_customers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/customers/{id}": {
      "get": {
        "operationId": "AdminController_customer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      },
      "patch": {
        "operationId": "AdminController_updateCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/inventory": {
      "get": {
        "operationId": "AdminController_inventory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/reviews": {
      "get": {
        "operationId": "AdminController_reviews",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/promotions": {
      "get": {
        "operationId": "AdminController_promotions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/content": {
      "get": {
        "operationId": "AdminController_content",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/rbac": {
      "get": {
        "operationId": "AdminController_rbac",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/rbac/roles/{id}": {
      "patch": {
        "operationId": "AdminController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/rbac/users/{id}": {
      "patch": {
        "operationId": "AdminController_updateUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "operationId": "AdminController_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      },
      "patch": {
        "operationId": "AdminController_updateSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
    "/api/v1/admin/audit-log": {
      "get": {
        "operationId": "AdminController_audit",
        "parameters": [
          {
            "name": "entity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin operations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Denes Commerce API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}